+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -125,7 +125,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-02-16T11:16:42.294Z",
+    "2026-02-16T14:11:04.151Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {