+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-16T11:16:38.883Z",
+    "2026-02-16T14:11:00.817Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {